ジャンプ兼決定ボタン割り当てをZ→SPACEキーに変更する準備。コメント解除すると可能

+++ クイズランリーグ/Source/Input/GamePad.h
@@ -1,6 +1,8 @@
 #pragma once
 
-using GamePadButton = unsigned int;
+//using GamePadButton = unsigned int;
+using GamePadButton = unsigned long long /*int*/;
+
 
 // ゲームパッド
 class GamePad
@@ -22,6 +24,24 @@ public:
 	static const GamePadButton BTN_RIGHT_SHOULDER	= (1 << 13);
 	static const GamePadButton BTN_LEFT_TRIGGER		= (1 << 14);
 	static const GamePadButton BTN_RIGHT_TRIGGER	= (1 << 15);
+
+	//static const GamePadButton BTN_UP               = (1 << 0ull);
+	//static const GamePadButton BTN_RIGHT            = (1 << 1ull);
+	//static const GamePadButton BTN_DOWN             = (1 << 2ull);
+	//static const GamePadButton BTN_LEFT             = (1 << 3ull);
+	//static const GamePadButton BTN_A                = (1 << 4ull);
+	//static const GamePadButton BTN_B                = (1 << 5ull);
+	//static const GamePadButton BTN_X                = (1 << 6ull);
+	//static const GamePadButton BTN_Y                = (1 << 7ull);
+	//static const GamePadButton BTN_START            = (1 << 8ull);
+	//static const GamePadButton BTN_BACK             = (1 << 9ull);
+	//static const GamePadButton BTN_LEFT_THUMB       = (1 << 10ull);
+	//static const GamePadButton BTN_RIGHT_THUMB      = (1 << 11ull);
+	//static const GamePadButton BTN_LEFT_SHOULDER    = (1 << 12ull);
+	//static const GamePadButton BTN_RIGHT_SHOULDER   = (1 << 13ull);
+	//static const GamePadButton BTN_LEFT_TRIGGER     = (1 << 14ull);
+	//static const GamePadButton BTN_RIGHT_TRIGGER    = (1 << 15ull);
+	//static const GamePadButton BTN_SPACE            = (1 << 16ull);
 
 public:
 	GamePad() {}
